Create README.md
Create a 3D illusion for a profile picture where a17 Year old cute Man  in 
a black hoodie Sitting casually on a Wingback chair. Wearing sneakers, and sunglasses, she looks ahead. The background 
features manibnd_ in big and capital Yellow neon light fonts on the dark 
grey wall. There should not be his shadow, and there are wings to make it 
appear as if he is an angel